float to double

+++ src/ArapDeformer.h
@@ -12,7 +12,7 @@ using namespace std;
 
 struct Constraint {
 	int vertexID;
-	Vector3f position;
+	Vector3d position;
 };
 
 class ArapDeformer {
@@ -21,8 +21,8 @@ public:
     ArapDeformer();
 	ArapDeformer(SimpleMesh *mesh);
     void initDeformation(vector<int> fixed_points);
-    void setHandleConstraint(int handleID, Vector3f newHandlePosition);
-    void applyDeformation(vector<int> fixed_points, int handleID, Vector3f handleNewPosition, int iterations);
+    void setHandleConstraint(int handleID, Vector3d newHandlePosition);
+    void applyDeformation(vector<int> fixed_points, int handleID, Vector3d handleNewPosition, int iterations);
     SimpleMesh m_mesh;
 
 private:
@@ -30,24 +30,24 @@ private:
     void estimateRotation();
     void updateB();
     void estimateVertices();
-    float calculateEnergy();
+    double calculateEnergy();
     void buildWeightMatrix(); 
     void calculateSystemMatrix();
-    Vector3f getConstraintI(int id);
+    Vector3d getConstraintI(int id);
     bool isInConstraints(int i);
     void updateSystemMatrix();
 
     //SimpleMesh m_mesh;
-    vector<MatrixXf> m_cell_rotations;
-    MatrixXf m_system_matrix;
-    MatrixXf m_system_matrix_original;
-    MatrixXf m_weight_matrix;
-    SparseMatrix<float> m_system_matrix_sparse;
+    vector<MatrixXd> m_cell_rotations;
+    MatrixXd m_system_matrix;
+    MatrixXd m_system_matrix_original;
+    MatrixXd m_weight_matrix;
+    SparseMatrix<double> m_system_matrix_sparse;
     int m_num_v;
     int m_num_p;
-    MatrixXf m_b;
+    MatrixXd m_b;
     int m_handle_id;
-    Vector3f m_new_handle_position;
+    Vector3d m_new_handle_position;
     vector<Constraint> m_constraints;
 
 };
